laser-equipment-intelligence/requirements.txt
# Core Framework
scrapy>=2.11.0
scrapy-playwright>=0.0.30
playwright>=1.40.0
twisted>=23.10.0

# HTML Parsing
beautifulsoup4>=4.12.0
parsel>=1.8.0
lxml>=4.9.0

# Anti-Detection & Evasion
undetected-playwright>=0.0.1
fake-useragent>=1.4.0
scrapy-impersonate>=0.0.1

# Proxy & CAPTCHA Services
requests>=2.31.0
2captcha-python>=1.1.3

# Database & Storage
psycopg2-binary>=2.9.0
redis>=5.0.0
sqlalchemy>=2.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
python-dateutil>=2.8.0

# OCR & PDF Processing
pytesseract>=0.3.10
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Geocoding & Location
geopy>=2.4.0
requests-oauthlib>=1.3.0

# Monitoring & Logging
prometheus-client>=0.19.0
python-slugify>=8.0.0
structlog>=23.2.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.0

# API Framework
flask>=3.0.0
flask-cors>=4.0.0

# PDF Processing & OCR
pytesseract>=0.3.10
PyMuPDF>=1.23.0
Pillow>=10.0.0

# LLM Integration
requests>=2.31.0

# ML/AI Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
beautifulsoup4>=4.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
httpretty>=1.1.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
pre-commit>=3.5.0
